Added stack data structure for parsing

+++ compiler/include/special_values.h
@@ -1,4 +1,4 @@
 #pragma once
 
-#define NOT_IN_A_DICTIONARY -1e9
-#define EXIT_REQUESTED      -1e8
+#define NOT_IN_A_DICTIONARY (int)-1e9
+#define EXIT_REQUESTED      (int)-1e8
